Add variant to hold differing PointCloud<PointT> types
Since we read a file with PointCloud2, which is agnostic to the file type but cannot be used
for algorithms, we can convert the PC2 into the actual PointCloud<PointT> type based on
the meta information read from the .pcd file into the PC2 object. Using this information we
can return the actual type, and use this variant for type agnostic access.

We will need to use std::visit and deduce the actial PointT type in order to use other
components of the library, which is a little cumbersome but I don't really know a better
way to do it.